+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 - Added simple .taf player using --play
 - Fixed certificate path issue on Windows-Integration for TeddyCloud upload
+- Added initial KDE integration for context menu handling
 
 ## [0.6.1] - 2025-05-24